Parametrized test_mode, cleaned up other funcs
